+++ include/game/entity.h
@@ -9,21 +9,25 @@ class Entity {
 protected:
     Game* game;
     Map* map;
-    H2DE_LevelPos pos;
     H2DE_LevelVelocity velocity = { 0.0f, 0.0f };
     EntityData data;
 
     H2DE_LevelObject* object = nullptr;
 
-//     H2DE_TimelineManager* tm = H2DE_CreateTimelineManager();
-
 //     Weapon* weapon = nullptr;
 //     std::vector<AutoWeapon*> autoWeapons;
-//     Face facing = LEFT;
+    H2DE_Face facing = H2DE_LEFT_FACE;
 //     Uint8 redFilterOpacity = 0;
 
+    H2DE_Surface* getSprite() const;
+    std::vector<H2DE_Hitbox> getHitboxes() const;
+
     virtual void updateImpl() = 0;
-//     void updateSprite();
+    void updatePos();
+    virtual void updateFacingImpl() = 0;
+    void updateFacing();
+    void updateAnimation(H2DE_LevelPos defaultPos);
+    void updateIndex();
 //     void updateWeapon();
 
     virtual void kill() = 0;
@@ -34,25 +38,25 @@ public:
 
     void update();
 
-//     void equipWeapon(int id); 
-//     void inflictDamages(float damages, float crit);
+    void inflictDamages(float damages, float crit);
+    //     void equipWeapon(int id); 
 
-//     H2DE_LevelPos getPos() const;
-//     EntityData getData() const;
+    EntityData getData() const;
+    H2DE_LevelObject* getObject() const;
+    H2DE_LevelObjectData* getObjectData() const;
+
 };
 
 
 
 class Player : public Entity {
 private:
-//     int xp = 0;
-//     int coins = 0;
+    int xp = 0;
+    int coins = 0;
 
     void updateImpl() override;
-//     void updateForControls();
-//     void updateForWorldCollisions();
-//     void updateAnimation(H2DE_LevelPos defaultPos);
-//     void updateFacing();
+    void updateForControls();
+    void updateFacingImpl() override;
 //     void updateForItems();
 //     void updateAutoWeapons();
 
@@ -73,14 +77,20 @@ public:
 
 class Enemy : public Entity {
 private:
+    bool canAttackNow = true;
+
     void updateImpl() override;
+    void updateFacingImpl() override;
 
     void kill() override;
-//     bool isNearPlayer();
+    bool isNearPlayer();
 
 public:
     Enemy(Game* game, Map* map, H2DE_LevelPos pos, EntityData data);
     ~Enemy();
+
+    void attacked();
+    bool canAttack() const;
 };
 
 #endif