+++ include/game/game.h
@@ -13,7 +13,13 @@ private:
     GameData* data = new GameData();
     Map* map;
 
+    std::vector<SDL_Keycode> keysDown = {};
+    std::vector<int> mouseButtonsDown = {};
+    H2DE_AbsPos mousePos = { 0, 0 };
+
     void initEngine();
+    void initMap();
+    void initCamera();
 
 public:
     Game();
@@ -26,6 +32,9 @@ public:
 
     H2DE_Engine* getEngine() const;
     GameData* getData() const;
+    std::vector<SDL_Keycode> getKeysDown() const;
+    std::vector<int> getMouseButtonsDown() const;
+    H2DE_AbsPos getMousePos() const;
 };
 
 #endif