+++ src/entities/enemy.cpp
@@ -15,7 +15,39 @@ void Enemy::kill() {
     
 }
 
+bool Enemy::isNearPlayer() {
+    static float nearPlayer = game->getData()->nearPlayer;
+    H2DE_LevelPos pos = getObjectData()->pos;
+    H2DE_LevelPos playerPos = map->getPlayer()->getObjectData()->pos;
+
+    return (std::abs(playerPos.x - pos.x) < nearPlayer && std::abs(playerPos.y - pos.y) < nearPlayer);
+}
+
+void Enemy::attacked() {
+    static int attackDelay = game->getData()->enemyAttackDelay;
+
+    canAttackNow = false;
+    H2DE_Engine::H2DE_Delay(attackDelay, [this]() {
+        canAttackNow = true;
+    });
+}
+
+bool Enemy::canAttack() const {
+    return canAttackNow;
+}
+
 // UPDATE
 void Enemy::updateImpl() {
-    
+    if (!isNearPlayer()) {
+        static H2DE_LevelObjectData* playerData = map->getPlayer()->getObjectData();
+        H2DE_LevelPos pos = getObjectData()->pos;
+        H2DE_LevelPos playerPos = playerData->pos;
+
+        float angle = std::atan2(playerPos.y - pos.y, playerPos.x - pos.x);
+        velocity = { data.stats.speed * std::cos(angle), data.stats.speed * std::sin(angle) };
+    } else velocity = { 0.0f, 0.0f };
+}
+
+void Enemy::updateFacingImpl() {
+    if (velocity.x != 0) facing = (velocity.x > 0) ? H2DE_RIGHT_FACE : H2DE_LEFT_FACE;
 }